Changes to Table types section to improve clarity
Adjusted Summary section to 15 tables instead of 13 as that's the count I think was intended.

Added abbreviated names of sections to table for clarity.

Adjusted each section so that the selector code was emphasized for clarity.

Adjusted section formatting in some sections to make it easier for a reader to figure out the tables vs informative descriptions of the tables.

Some small typos/wording changes.

Split out count codes into a list for clarity.