Split out list from version sections for clarity
Split out the list of elements of the version string sections for clarity of how they're alike and how they're different.

Added an equivalent calculation to make it clear where the 2^24 computation was coming from in those sections.

Demoted Version 2.XX section to match 1.XX section level to make it conceptually clear that both of these version strings are on the same conceptual level.